Circulars osb side (#5351)
