+++ src/lib/util/smallUtils.ts
@@ -8,7 +8,6 @@ import { Bank, Items } from 'oldschooljs';
 import { ItemBank } from 'oldschooljs/dist/meta/types';
 import { MersenneTwister19937, shuffle } from 'random-js';
 
-import { ClueTiers } from '../clues/clueTiers';
 import { skillEmoji } from '../data/emojis';
 import type { ArrayItemsResolved, Skills } from '../types';
 import getOSItem from './getOSItem';
@@ -151,23 +150,6 @@ export function makeEasierFarmingContractButton() {
 		.setLabel('Ask for easier Contract')
 		.setStyle(ButtonStyle.Secondary)
 		.setEmoji('977410792754413668');
-}
-
-export function buildClueButtons(loot: Bank | null, perkTier: number) {
-	const components: ButtonBuilder[] = [];
-	if (loot && perkTier > 1) {
-		const clueReceived = ClueTiers.filter(tier => loot.amount(tier.scrollID) > 0);
-		components.push(
-			...clueReceived.map(clue =>
-				new ButtonBuilder()
-					.setCustomId(`DO_${clue.name.toUpperCase()}_CLUE`)
-					.setLabel(`Do ${clue.name} Clue`)
-					.setStyle(ButtonStyle.Secondary)
-					.setEmoji('365003979840552960')
-			)
-		);
-	}
-	return components;
 }
 
 export function makeAutoFarmButton() {
